Revise host-spanning check.
Previously, we used the --domains option of wget to crawl example.com
and www.example.com.

Except that's not how wget interpreted the request.  It interpreted it
as sites _belonging to_ those domains, so e.g. you'd crawl
foo.example.com, bar.example.com, ...

That isn't quite what we wanted for the www = bare domain heuristic
(which, for better or worse, is in wide use on the Web), so that
heuristic has been re-coded in Lua.