Log errors; accept wget result on any error.
Wget errors happen.  However, instead of aborting the fetch, we now
accept any exit code generated by wget[1] and log the errors.

The logging is a bit verbose, as it will log things that wget reports as
errors but aren't really errors.  (Redirections, for example.)  That
will be tuned later.

[1]: SIGSEGV and friends still generate failure conditions, which need
to be somehow handled.