Record *all* wget retrievals to Redis.
Determining what's an error and what isn't can be a bit messy.  For
example, wget returns NEWLOCATION for 3xx responses, which is
technically not a RETRFINISHED -- but wget can retry the fetch on the
new URL and will pass.

However, under the old scheme, we'd still log an error.  That doesn't
seem right.

In an attempt to keep archivebot.lua focused solely on data gathering
and fetching, I am moving the error-processing code to the bot frontend.

One nice side benefit: we can remove the crazy
redirect-stdout-to-Redis-writer hack in the Seesaw pipeline.