Rework Job as a cached-data container with some mutators.
This is one step to building a web dashboard for ArchiveBot.

I decided to do this because a dashboard is going to require most of the
Redis data for a job, and multiple HGETs is much less efficient than one
(larger but less per-command latency) HGETALL.

Neat side effect: LogAnalyzer is WAY shorter.  It is a bit less
efficient in terms of Redis queries, but we'll just have to see how bad
that really is.  I suspect the answer is "not at all".