Use sorted sets for archive job logs.
Although insertion is slower, the error analysis process needs to pull a
range of log entries.  With lists, this can be (and often is) O(n),
where n is the number of entries in the list.  With sorted sets, it's
O(log(n)+m), where n is the number of entries in the set and m is the
number of elements that will be considered.  (Usually, n >>> m.)

This commit also eliminates the Lua-based URL categorization, because
it's no longer needed.