Record jobs in a CouchDB database.
Why CouchDB and not Redis?

Redis keeps its whole dataset in RAM.  This is fine for datasets of
bounded size or datasets that will likely not exceed available RAM, but
not so good for datasets that (1) grow without bound and (2) absolutely
need to stay around.

So we use CouchDB for cold storage.