Stop failing at dating.
Well, maybe.

So:

1. Time#to_i returns seconds since the UNIX epoch.
2. new Date(number) takes milliseconds since the UNIX epoch.

No offset arithmetic is necessary.

To cleanse my hands of all of this annoying date stuff in the dashboard,
I'm dragging in moment.js.  It's not too big and it makes my intentions
clearer.  (So, if I've still goofed, at least someone can see what I
meant to do.)