Move logging to a script.
I want to display upload status in the dashboard.  The easiest way to do
this is to capture stdout, as that's where rsync progress is displayed.

To do this, we'll need to generate log entries from the wget Lua script
*and* the pipeline.  Easiest way to commonize the behavior: move it to
Redis.

As part of this commit, I've introduced a caching script executor in
Lua.