Add a heartbeat to each work item.
The heartbeat starts once the item's been pulled out of the queue, and
stops right before the item is marked as done.

The intent of this system is to provide automatic detection of job (or
pipeline) failure, which has become an issue now that we are using
multiple fetch machines.