Anonymize pipeline identity.
We will soon display pipelines for each job in the dashboard.  The
pipeline ID is constructed from potentially sensitive information: the
FQDN, hostname, and process ID.  Best to obscure it.

MD5 was chosen because it has the shortest hex digest of all hashes in
hashlib, and we do not need a collision-resistant hash function for this
purpose -- just something to obscure the input.