Bump death threshold to 3600 cycles.
Rationale:

    [00:56:35] <ivan`> how about more like 2 hours to deal with
                       overloaded VMs and migrating VMs?

Assuming continuous reaper operation, this grants a ~2 hour grace period
for jobs in limbo.

Bumping the threshold up this high also means that some sort of manual
reap command is needed for the scenario where we know a job isn't coming
back and don't want to wait for the reaper to do its thing.