Ignore set registration shouldn't crap out if there's no ignore sets.
Brain#add_ignore_sets should always be called with zero or more ignore
sets.  It's possible, however, that params['ignore_sets'] could be nil
(i.e. if it wasn't set at all).

Arguably one could put this check in the caller or the callee.  The
caller seems more right to me, which is weird, since I'm not from New
Jersey.  Meh.  I'll fix it later if it ends up being a problem.