Always supervise and register a TwitterTweeter.
This eliminates the need to pass opts into Broadcaster, and simplifies
cog initialization.  It does cost a couple of branches inside
TwitterTweeter, but I think we can live with those.

This commit also imposes additional restrictions on TwitterTweeter's
public interface to demonstrate that TwitterTweeter really does in fact
do nothing if it isn't configured.