Propagate PATH to wpull invocation.
Motivation: It can be useful to run ArchiveBot's Python code in a
virtualenv.  Two examples:

1. Many Ubuntu systems (still) don't use Python 3 as their default
   Python environment.  Python 3 is present, but it is easier (for me)
   to virtualenv it rather than do a bunch of PATH magic to get stuff
   set up right.

2. Virtualenv could be used to write a one-command-does-all script for
   setting up an ArchiveBot pipeline.

However, the wpull invocation from the pipeline sets an entirely new
process environment that doesn't include the virtualenv-modified PATH.
The wpull wrapper script then tries to find the wpull package, fails,
and all hell breaks loose.

This commit fixes the above problem by propagating the pipeline's PATH
to the wpull environment.  This is a bit overly broad (really, we just
need the directory of the Python executable used to run the pipeline)
but I think it should be okay for now.  Security gurus are welcome to
tear this decision apart.