Cut log scrollback in half.
ArchiveBot is currently running with 31 concurrent jobs.  This is more
than has come before, but it is arguably a reasonable concurrency level.
With 512 entries per backlog, the dashboard eats up around a gigabyte of
RAM in Firefox and Chrome.  Hopefully, cutting down the scrollback will
bring that level down a bit.