Reduce LogView memory requirements.
Previously, a LogView consisted of many LogEntryViews, and each
LogEntryView was pretty heavy: it wasn't unusual to have the dashboard
sucking up a gigabyte of RAM (immediately!) with a couple thousand log
entries in play.

This commit greatly reduces the complexity of each LogEntryView, and in
doing so appears to greatly reduce memory consumption.  Anecdote: using
Firefox 27.0.1 (Canonical's build), I can view 3,930 log entries across
17 jobs comfortably.  After running Firefox's GC and CC, about:memory
claims that the dashboard uses 200 MB, which is still pretty fat but
much less so.