Switch to bleeding-edge redis-py.
Andy McCurdy has merged a large number of changes to redis-py's pubsub
implementation.  Of particular interest is pubsub.run_in_thread(), which
subscribes to a pubsub channel on a separate thread.  In released
versions of redis-py, doing this is very painful.

These changes are scheduled for redis-py 2.10.0, which is not yet
released.  Hence this commit.

We also drag in hiredis because redis-py edge + native Python parser
gets into infinite loops on sudden connection failure in the pubsub
client.  With hiredis, a redis.exceptions.ConnectionError is raised,
which is much easier to work with.