Use hiredis for the dashboard's Redis connection.
hiredis is recommended for use when the Redis connection may deal with
large datasets, e.g. results from ZRANGE* queries.  That describes the
dashboard pretty well (i.e. Job#read_new_entries) so we might as well
give it a shot.

The rationale is "performance", but I wonder if that also includes
"memory usage".