Introduce !ao-only mode for pipelines.  Closes #44.
In !ao-only mode, a pipeline listens only to the pending-ao queue and
its private queue.[1]

This commit also changes the bot so that all !ao jobs are put onto a
separate queue.  This lets us use the same queuing behavior that we use
for the existing queue.

[1] A pipeline in "!ao-only" mode can still do !a jobs if they're
directed specifically to it.  This is a violation of the strict reading
of "!ao-only", but pipeline-specific queues have been used in the past
to test tweaks, and I think that's a useful use case to keep around.
We'll see in whether this exception proves to be too much of a pain.