Ignore pattern reports may now be suppressed or shown.
Suppressing ignores can be useful when a job has many matching URLs.  In
this situation, reporting every ignore:

1. results in quite a bit of network traffic
2. causes the job log to scroll too fast for most humans to understand

In cases like this, the log becomes more useful when ignore pattern
matches are not logged.  The old "show ignores" function attempted to do
this, but ignores were still present in the dashboard log -- they were
just CSS-hidden.  As a result, it was possible for a large stream of
ignores to boot useful log data out of the job's log buffer.

Suppressing ignores can also improve job speed: there's a small but
significant amount of time eaten up by ZADDing an ignore message to the
job log and waiting for a response.