Dashboard: eliminate pipelines set and PublicPipelineRecord.
We now SCAN for keys matching pipeline:* and just return the result of
HGETALL on those keys.  This makes loading the pipeline report slower,
but it also means that we don't block Redis for an extended period of
time with potentially huge SMEMBERS/KEYS calls.

This commit also eliminates PublicPipelineRecord because hiding that
information hurts pipeline operators more than it helps them.