plumbing: Introduce set-watermarks.
The set-watermarks program watches the number of URLs queued (note: the
total number of URLs queued, not the current queue length).  When the
number of URLs queued exceeds a multiple of 100000, the next multiple of
100000 is set as the next watermark.

If the number of URLs queued equals the watermark, we can use that state
to send out notifications or some such.