bot: Experiment with finer !con permissions.
I think that raising job concurrency is something that needs to be
restricted to channel operators (which usually implies a more trusted
set of people).  Reasons:

1. There's quite a number of sites out there that will block for
   aggressive crawling.
2. In some cases (say, where multiple wpulls have to share a core)
   running one job with high concurrency can adversely affect other jobs
   running on the same pipeline.

This is the first fine-grained restriction that has made it to
ArchiveBot and I'm not sure if it's a good idea.  Let's see what
happens.