doc: Add ZeroMQ 4.0.5 to backend requirements.
I'm not actually sure of the best way to get ZeroMQ, so I've left it
out.  Ubuntu Utopic seems to still be on ZeroMQ 2.2; other systems are
on 4.0.5.  (For my development system, I just built ZeroMQ 4.0.5 from
the tarball.)  It'd be nice to have e.g. a PPA to point to.