bot: Make !a operate in no-offsite-links by default.
A major job bloating factor is offsite links and their associated page
requisites.  This extra fetch work and pipeline load should be a
conscious decision, not the default.

It would be cleaner to have --offsite-links write "offsite_links" into
the job record, but doing that would require a pipeline upgrade.  This
doesn't.