Eliminate abort_requested flag.  #136.
The difference between abort_requested and aborted is when the Wpull
hooks get around to checking for the former in order to set the latter.
In practice, however, this distinction hasn't been all that useful.

Having a single aborted flag makes it easier to answer the question "is
this job aborted?" from _anywhere_ in the pipeline.