Check for job abort at WgetDownload (now WpullDownload) start.  #136.
An indefinite number of retries can be useful if a pipeline operator is
tinkering with a wpull process via e.g. manhole or kill -{STOP,CONT} to
update code, unwedge it, etc.  (These should be much less frequent
scenarios in wpull 1.0, but could still happen.)

Such tinkering may crash the wpull process.  This is okay for wpull,
because it can resume using --database.  Indefinite retries means that
crashed jobs will be able to resume where they let off.

---

There are still some situations where the wpull process is in a state
that cannot be feasibly fixed.  In these situations, we want to be able
to abort the job.

The current wpull hooks do an abort check after processing a response,
so a wpull restart should be able to pick up most aborts.  However, if
we have a situation where wpull cannot even start or finish one response
[1], abort will not take effect.  Abort processing in a pipeline task
means that we can handle aborts in this case.

[1]: AFAIK, the preflight check is designed to address all known
     situations like this.  However, doing the abort check from both the
     pipeline and wpull hooks can be done without much code duplication,
     so this is probably fine.