bot: Disallow recursive youtube-dl.  #82.
A test job on Youtube[1] had youtube-dl crawling channel pages and video
pages; it ended up downloading videos multiple times.  Until this
behavior can be prevented, I think the best thing to do is to not allow
recursive retrieval.  A workaround (albeit tedious) is to identify all
pages in a job that have an embedded video, generate a list of URLs, and
use !ao < URL.

[1]:

json:
https://archive.org/download/archiveteam_archivebot_go_20150513040003/www.youtube.com-inf-20150512-214850-4jr0n-aborted.json

warc.gz:
https://archive.org/download/archiveteam_archivebot_go_20150513040003/www.youtube.com-inf-20150512-214850-4jr0n-00000.warc.gz