Added additional ignore patterns for Russian MediaWiki installations.
Ignore patterns for Russian MediaWiki installations. I’m not actually
sure if this (“Uchastnik_blog”) is how MediaWiki localizes “User_blog”
into Russian but I figure it's as good a guess as any. I disclaim
responsibility if these are incorrect, and offer this with no warranty
or guarantee of accuracy….. I found them by using the URLs that the
English equivalents of these pages redirected to on wiki.rutracker.org.