ops: Use docker unless-stopped restart policy for ArchiveBot containers.
A popular update strategy used for host OSes is to restart-on-update.
This assumes container state is persisted someplace safe and that they
come back on system start.

It should be fine to restart ArchiveBot backend services, but we need
them to automatically come online again.  The unless-stopped policy will
do this for us.

[ci skip]