dashboard3: Remove Safari-specific onmousewheel case.
https://github.com/HaxeFoundation/haxe/issues/4005 mentions that the
onmousewheel export was removed from Haxe 3.2 because onwheel is the new
standard name.  Ubuntu 16.04 includes Haxe 3.2.1.

To prepare the dashboards for HTTPS, I need to make some source
modifications.  In the case of dashboard3, this means recompiling the
Haxe code, which means I need to be able to compile it as-is before
making modifications.