bot: remove broken smarts in !status URL.
The bot used to contact a CouchDB database of previously-run jobs, but
this database has turned out to be more of a pain than it is worth.
The IA ArchiveBot collections are the only real source of truth for this
data and we already have a viewer for that.

It is possible to add an API endpoint to the viewer to query a host (and
perhaps eventually full URIs), but I don't feel like doing that right
now.