New dashboard WebSocket server
The old server was unable to keep up with the messages (#333), ate all the CPU and RAM it could get (#324), got unresponsive occasionally (killer.py), etc.

The new server will drop messages when it can't send them to the client quickly enough (backpressure). The lack of this measure is probably what caused the high memory usage in the old server.

The new server runs on port 4568 instead of 4567 because the latter is occupied by the old server (serving the dashboard page etc.) on the EFNet setup.