Handle the data directory not existing
The data directory is created by seesaw when it starts running the Pipeline (or specifically, the first item through the pipeline). The only reason this didn't already crash prior to 1585458a is that the Tornado IOLoop which runs the monitoring apparently only gets launched from within seesaw and only runs the report() *after* seesaw started processing the first item.