Include CentralAuth to mediawiki ignore rule
On wikimedia wikis, CentralAuth is the pandora's box responsible for
crazy amount of other WMF-wiki userpages. See mine [1] for example -
it has 775*3 links to WMF-sisterprojects as of 2019-12-26, and it only
increases when the new wiki is added to WMF farm. Therefore it is
sensible to remove them from the list.

[1]: https://w.wiki/EUG

Signed-off-by: Yongmin Hong <revi@pobox.com>