Add Google and Outlook calendars to global ignoreset
These are often generated by the Tribe events calendar on wordpress sites, and always redirect to a generic login page (or an information page for the outlook.live.com variant). These pages aren't useful to save (especially as even if you have an account on one of these services, loading the login page on web.archive.org won't be helpful).