When a new job is added, don't scroll the stopped logged window being hovered over
This fixes:

"when you hover over a log window, it stops scrolling as intended, but
each time a new job is added at the top of the dashboard, the log window
scrolls to the bottom. I'm pretty sure this wasn't the case before [...]."

Diagnosis:

`JobsRenderer._createLogContainer` calls `this.applyFilter()`; `applyFilter`
was scrolling all the log windows (with the intention of scrolling log windows
made newly visible by the filter), but neglected to check whether they were
`log-window-stopped`.