Add Wix ignore set
This has been in use for a while, and generally works well. Wix likes to give 200s for URLs that should be 404s (and gives a 404 message on those pages too).

Wix sometimes manages to find other ways of looping (sometimes repeating _404_ in the URL, potentially without a / between them), but that is comparatively rare and I'm not aware of a nice ignore for it (I generally ignore _404_.*_404 in that case, but that only makes sense after having seen it already starting to loop).